+++ docs_src/models/constraints.py
@@ -1,0 +1,17 @@
+import edgy
+import sqlalchemy
+from edgy import Database, Registry
+
+
+database = Database("sqlite:///db.sqlite")
+models = Registry(database=database)
+
+
+class User(edgy.Model):
+    name = edgy.fields.CharField(max_length=255)
+    is_admin = edgy.fields.BooleanField()
+    age = edgy.IntegerField(null=True)
+
+    class Meta:
+        registry = models
+        constraints = [sqlalchemy.CheckConstraint("age > 13 OR is_admin = true", name="user_age")]